Fix when Helm provider ignores FAILED release state
This moves status field to non-computed field in order to track state of
Helm release and enforce that the state is `DEPLOYED`.

This fixes behaviour when Terraform would ignore previously created
release that is in FAILED state and failed to try to bring it to desired
state (assuming this will always be `DEPLOYED`)

See
https://github.com/terraform-providers/terraform-provider-helm/issues/159
for more detail.